Reorganise main gui CMakeLists.txt and remove first hui file.
