+++ src/gui/addForcefieldTermsDialogFuncs.cpp
@@ -1,25 +1,20 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2024 Team Dissolve and contributors
 
-#include "classes/atomType.h"
 #include "classes/species.h"
 #include "gui/addForcefieldTermsDialog.h"
 #include "gui/models/addForcefieldDialogModel.h"
-#include "templates/algorithms.h"
-#include <QInputDialog>
-#include <QMessageBox>
+#include <QHBoxLayout>
 #include <QQmlContext>
 #include <QQuickItem>
 #include <QQuickWidget>
-#include <set>
 
 AddForcefieldTermsDialog::AddForcefieldTermsDialog(QWidget *parent, Dissolve &dissolve, Species *sp) : QDialog(parent)
 {
-
     // For the QIcon theme
     // QIcon::setThemeName("personal");
 
-    QQuickWidget *view = new QQuickWidget(QUrl("qrc:/dialogs/qml/AddForcefieldTermsDialog.qml"), this);
+    auto *view = new QQuickWidget(QUrl("qrc:/dialogs/qml/AddForcefieldTermsDialog.qml"), this);
     AtomTypeModel atModel(dissolve.coreData());
 
     atModel.setData(dissolve.coreData().atomTypes());
@@ -38,9 +33,7 @@ AddForcefieldTermsDialog::AddForcefieldTermsDialog(QWidget *parent, Dissolve &di
     connect(model, SIGNAL(cancel()), this, SLOT(reject()));
     connect(model, SIGNAL(accept()), this, SLOT(accept()));
 
-    QHBoxLayout *topLeftLayout = new QHBoxLayout;
+    auto *topLeftLayout = new QHBoxLayout;
     topLeftLayout->addWidget(view);
     setLayout(topLeftLayout);
-
-    // ui->verticalLayout->addWidget(container);
 }
