+++ src/gui/mainTabsWidget.h
