+++ src/gui/thread.h
